move presenter out of folder
